server/utils/requirements.txt
face_recognition==1.3.0
firebase_admin==6.5.0
numpy==1.26.4
Pillow==10.3.0
Requests==2.32.2
face_recognition_models==0.3.0
face_recognition==1.3.0
git+https://github.com/ageitgey/face_recognition_models